Make mousewheel work on sliders
